+++ ChangeLog.md
@@ -7,6 +7,7 @@ Compression streams ChangeLog
 
 * Merged PR #8: Add ZStandard compression, based on the `zstd` extension
   from https://github.com/kjdev/php-ext-zstd/ - see issue #7.
+  (@thekid)
 * **Heads up:** Dropped support for PHP < 7.4, see xp-framework/rfc#343
   (@thekid)
 * **Heads up:** Algorithm implementations must change their `compress`